requirements.txt
Flask==3.0.0
pymongo==4.5.0
joblib==1.3.2
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.0
requests==2.31.0
flask-caching==2.1.0
apscheduler==3.10.4
python-dotenv==1.0.0